+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Improved update-lambdas to ensure zips get built
+
 ### Changed
 
 - Moved modules to a /modules folder
@@ -17,11 +19,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 
+
+
+
 ## [1.0.0] - 2025-10-30
 
 ### Added
-
-- Improved update-lambdas to ensure zips get built
 
 - Support inputs.stac_server_version usage